test: Extract Radio.Configuration.Tests from scattered test projects
Move 115 configuration tests into a dedicated Radio.Configuration.Tests project:
- 12 tests from Radio.Core.Tests/Configuration/ (DatabaseOptions, PathResolver)
- 77 tests from Radio.Infrastructure.Tests/Configuration/ (stores, secrets, backup, bridge)
- 11 tests from Radio.IntegrationTests/Secrets/ (secrets integration)

Organized into subdirectories: Backup/, Bridge/, Integration/, Models/, Options/, Secrets/, Stores/
