+++ RadioConsole.sln
@@ -51,6 +51,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Radio.Fingerprinting", "src
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Radio.Metrics.Tests", "tests\Radio.Metrics.Tests\Radio.Metrics.Tests.csproj", "{60DB543F-7498-4404-A9B5-E09988B60F9C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Radio.Configuration.Tests", "tests\Radio.Configuration.Tests\Radio.Configuration.Tests.csproj", "{22F2CDB8-4450-4C72-8667-80BB20F60B3F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -313,6 +315,18 @@ Global
 		{60DB543F-7498-4404-A9B5-E09988B60F9C}.Release|x64.Build.0 = Release|Any CPU
 		{60DB543F-7498-4404-A9B5-E09988B60F9C}.Release|x86.ActiveCfg = Release|Any CPU
 		{60DB543F-7498-4404-A9B5-E09988B60F9C}.Release|x86.Build.0 = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|x64.Build.0 = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Debug|x86.Build.0 = Debug|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|x64.ActiveCfg = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|x64.Build.0 = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|x86.ActiveCfg = Release|Any CPU
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -339,5 +353,6 @@ Global
 		{928DC03B-E6A1-4E4D-8D2C-6B9494831D35} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{82818C12-06D6-49A3-885D-C2DF885FC94B} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{60DB543F-7498-4404-A9B5-E09988B60F9C} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{22F2CDB8-4450-4C72-8667-80BB20F60B3F} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 EndGlobal
