+++ tests/Radio.Configuration.Tests/Bridge/SqliteConfigurationProviderTests.cs
@@ -1,4 +1,4 @@
-namespace Radio.Infrastructure.Tests.Configuration;
+namespace Radio.Configuration.Tests.Bridge;
 
 using Microsoft.Data.Sqlite;
 using Microsoft.Extensions.Configuration;
